Add babel-jest into the repo
Summary:babel-jest is moving into jest as a package. This allows us to consolidate the codebase. Note that we are only re-using the npm package for babel-jest but the implementation actually changed as the `babel-preset-jest` is now being loaded automatically.

When `babel-jest` is installed in the project that jest is being run on, it will now automatically be loaded. See the `examples/react` folder – the only setup needed for babel to work is to enable the default presets (es2015 and react) and to `npm install babel-jest`. In this PR I also added the `setupFiles` config flag. There isn't really a reason to limit this to a single entry.
Closes https://github.com/facebook/jest/pull/757

Reviewed By: kentaromiura

Differential Revision: D2998065

fb-gh-sync-id: D2998065
shipit-source-id: D2998065