babel-jest should compile react-native.
Summary:This changes it so react-native always gets compiled properly.

I thought long and hard about whether there should be an explicit whitelist, and yeah, probably there should be, but I don't want to overcomplicate things for now. One of the goals of Jest is to simplify the setup process, especially around Facebook's open source projects, and this helps greatly. New documentation for react-native testing is coming probably tomorrow.

see #770.
Closes https://github.com/facebook/jest/pull/774

Differential Revision: D3017346

fb-gh-sync-id: e0c1444ce7c247fa3187ae024b38c81d19035f54
shipit-source-id: e0c1444ce7c247fa3187ae024b38c81d19035f54